scripts/xml_to_ods/requirements.txt
lxml
odsgenerator
progressbar2